pbuilder-hooks: Only pass one changes file to lintian
Now that we generate both a binary and source .changes files,
we need to take special care to not pass them both to
lintian. Try to avoid passing the _source.changes file to lintian
(as that will only give us partial lintian results, not checking
binary packages obviously) and also try to avoid hard-coding
amd64 as the architecture for binary build.