pbuilder-hooks: Disable list-missing hook
It's written in an outdated way and nowadays dh_missing should
do this job and use list-missing by default in newer compats.
(Hopefully switching to fail-missing in a not to distant future
compat.)

Rather than updating the hook code, just disable it for now and
rely on dh_missing becoming ever more prevalent.